updates project structure for 1.x release
