+++ .github/workflows/main.yml
@@ -13,31 +13,34 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.8", "3.9", "3.10", "3.11"]
+        python-version: ["3.10", "3.11", "3.12"]
     steps:
-    - uses: actions/checkout@v3
-    - name: Configura o Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v4
-      with:
-        python-version: ${{ matrix.python-version }}
-    - name: Instala os requisitos do módulo
-      run: |
-        python -m pip install -r dev-requirements.txt
-    - name: Executa os testes do pacote
-      run: |
-        python -m pytest --cov-report=xml --cov=sintetizador tests/
-    - uses: codecov/codecov-action@v1
-      with:
-        token: ${{ secrets.CODECOV_TOKEN }} 
-        files: ./coverage.xml
-        flags: unittests
-        env_vars: OS,PYTHON
-        name: codecov-sintetizador
-        fail_ci_if_error: true 
-        verbose: true 
-    - name: Veriricações de qualidade de código - tipagem estática
-      run: |
-        python -m mypy ./sintetizador
-    - name: Verificações de qualidade de código - linter PEP8
-      run: |
-        python -m pylama --ignore E203 ./sintetizador
+      - uses: actions/checkout@v4
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
+      - name: Set up Python ${{ matrix.python-version }}
+        run: uv python install ${{ matrix.python-version }}
+      - name: Install the project
+        run: |
+          uv sync --all-extras --dev
+      - name: Runs tests
+        run: |
+          uv run pytest  --cov-report=xml --cov=app ./tests
+      - uses: codecov/codecov-action@v4
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: ./coverage.xml
+          flags: unittests
+          env_vars: OS,PYTHON
+          name: codecov-sintetizador
+          fail_ci_if_error: true
+          verbose: true
+      - name: Static type check
+        run: |
+          uv run mypy ./app
+      - name: Linter code check
+        run: |
+          uv run ruff check ./app
+      - name: Sphinx build
+        run: |
+          uv run sphinx-build -M html docs/source docs/build
