+++ .gitignore
@@ -161,3 +161,7 @@ cython_debug/
 
 deck_teste/*
 docs/build/**
+
+# UV
+
+uv.lock
