+++ app/model/operation/spatialresolution.py
