+++ app/utils/terminal.py
@@ -1,0 +1,49 @@
+import asyncio
+from typing import List, Optional, Tuple
+
+RETRY_DEFAULT = 3
+TIMEOUT_DEFAULT = 10
+
+
+async def run_terminal_retry(
+    cmds: List[str],
+    num_retry: int = RETRY_DEFAULT,
+    timeout: float = TIMEOUT_DEFAULT,
+) -> Tuple[int, str]:
+    """
+    Runs a command on the terminal (with retries) and returns.
+
+    :param cmds: Commands and args to be executed
+    :param num_retry: Max number of retries
+    :param timeout: Timeout for giving up on the command
+    :return: Return code and outputs
+    :rtype: tuple[int | None, str]
+    """
+    for _ in range(num_retry):
+        cod, outputs = await run_terminal(cmds, timeout)
+        if cod == 0:
+            return cod, outputs
+    return -1, ""
+
+
+async def run_terminal(
+    cmds: List[str], timeout: float = TIMEOUT_DEFAULT
+) -> Tuple[Optional[int], str]:
+    """
+    Runs a command on the terminal and returns.
+
+    :param cmds: Commands and args to be executed
+    :param timeout: Timeout for giving up on the command
+    :return: Return code and outputs
+    :rtype: tuple[int | None, str]
+    """
+    cmd = " ".join(cmds)
+    proc = await asyncio.create_subprocess_shell(
+        cmd, stdout=asyncio.subprocess.PIPE, stderr=asyncio.subprocess.PIPE
+    )
+    stdout, stderr = await asyncio.wait_for(proc.communicate(), timeout=timeout)
+    if stdout:
+        return proc.returncode, stdout.decode("utf-8")
+    if stderr:
+        return proc.returncode, stderr.decode("utf-8")
+    return None, ""
