--- dev-requirements.txt
@@ -1,14 +1,0 @@
-click
-pyarrow
-idessem>=0.0.16
-pylama
-mypy
-pytest
-pytest-cov
-sphinx
-sphinx-gallery
-sphinx-rtd-theme
-plotly
-numpydoc
-nb2plots
-black
