+++ pyproject.toml
@@ -1,0 +1,61 @@
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "sintetizador-dessem"
+dynamic = ["version"]
+dependencies = [
+    "click>=8.1.8",
+    "idessem>=0.0.23",
+    "pyarrow>=19.0.0",
+]
+requires-python = ">= 3.10"
+authors = [
+  {name = "Rogerio Alves", email = "rogerioalves.ee@gmail.com"},
+  {name = "Mariana Noel", email = "marianasimoesnoel@gmail.com"},
+]
+description = "sintetizador-dessem"
+readme = "README.md"
+license = {file = "LICENSE"}
+classifiers = [
+    "Development Status :: 4 - Beta",
+    "Intended Audience :: Developers",
+    "License :: OSI Approved :: MIT License",
+    "Programming Language :: Python :: 3.10",
+    "Operating System :: OS Independent",
+]
+
+[project.optional-dependencies]
+dev = [
+    "pytest",
+    "pytest-cov",
+    "ruff",
+    "mypy",
+    "sphinx-rtd-theme",
+    "sphinx-gallery",
+    "sphinx",
+    "numpydoc",
+    "plotly",
+    "matplotlib",
+]
+
+
+[project.urls]
+Documentation = "https://rjmalves.github.io/sintetizador-dessem/"
+Repository = "https://github.com/rjmalves/sintetizador-dessem/"
+
+[project.scripts]
+sintetizador-dessem = "main:main"
+
+[tool.hatch.version]
+path = "app/__init__.py"
+
+[tool.hatch.build.targets.wheel]
+include = [
+  "main.py",
+  "app/",
+]
+
+[tool.ruff]
+line-length = 80
